services/appointment_service/requirements.txt
flask
flask-restx
flask-cors
uvicorn
pymongo
python-dotenv
psycopg2-binary
pydantic[email]
fastapi-utils
typing-inspect
httpx